+++ internal/endtoend/testdata/each/schema.sql
@@ -1,0 +1,5 @@
+CREATE TABLE devices (
+    id Uint64 NOT NULL,
+    name Utf8,
+    PRIMARY KEY (id)
+);